+++ src/evm/config/tokens/token-maps/stables.ts
@@ -39,6 +39,7 @@ import {
 import { EvmToken } from '../../../currency/token.js'
 import { AUSD } from '../tokens/AUSD.js'
 import { USDT0 } from '../tokens/USDT0.js'
+import { HONEY } from '../tokens/berachain/HONEY.js'
 
 export const STABLES = {
   [EvmChainId.ARBITRUM]: [
@@ -243,7 +244,11 @@ export const STABLES = {
     AUSD[EvmChainId.KATANA],
   ],
   [EvmChainId.HYPEREVM]: [USDT[EvmChainId.HYPEREVM]],
-  [EvmChainId.BERACHAIN]: [USDT[EvmChainId.BERACHAIN]],
+  [EvmChainId.BERACHAIN]: [
+    USDT[EvmChainId.BERACHAIN],
+    USDC[EvmChainId.BERACHAIN],
+    HONEY[EvmChainId.BERACHAIN],
+  ],
   // TESTNETS
   [EvmChainId.ARBITRUM_SEPOLIA]: [],
   [EvmChainId.SEPOLIA]: [],